src/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
networkx>=3.1
requests>=2.31.0
pydantic>=2.4.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0